alloctest: A simple malloc() like circular buffer
A tiny example in C of a circular buffer malloc()/Free() like implementation.

It assumes that the memory allocated is freed in the approximate same order it is allocated, so that it circles around. Memory must not strictly be freed in the order like a queue would be. But it can't be allocated until the first element is free again.

It shows some unit test cases that handles wrapping and when buffers are full.

The code is not thread-safe. When allocating, the head and the block must be updated atomically (with any other alloc/free). Likewise, when freeing, the tail and the block must be updated atomically (with any other alloc/free).